Fix LOCALCONF_VERSION version check in distro/e-ale.conf
We intentionally bumped the configuration version in our
local.conf.sample template, but we also need to let Bitbake
know what version we are expecting to see. This can be done in
a sanity.conf file like in poky, but we'll just put it in
our distro/e-ale.conf file.